% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_pdf_png}
\alias{convert_pdf_png}
\title{Convert a PDF file to a series of PNG image files}
\usage{
convert_pdf_png(path, verbose = TRUE, dpi = 600)
}
\arguments{
\item{path}{Path to the PDF file that needs to be converted to PNGs.}

\item{verbose}{A logical value indicating whether to display progress
messages during the conversion process. The default value is TRUE}

\item{dpi}{The resolution in dots per inch (dpi) to be used for the PNG
images. The default value is 600.}
}
\description{
Uses \code{pdftools::pdf_convert()} for conversion.
}
