% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_pptx_pdf}
\alias{convert_pptx_pdf}
\title{Convert a PPTX file to a PDF file}
\usage{
convert_pptx_pdf(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to the PPTX file that needs to be converted to PDF.}

\item{verbose}{A logical value indicating whether to display progress
messages during the conversion process. The default value is TRUE}
}
\description{
Uses \code{docxtractr::convert_to_pdf()} for conversion.
}
