% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_notes_xml}
\alias{extract_notes_xml}
\title{Extract Notes from XML}
\usage{
extract_notes_xml(file, collapse_text = TRUE, xpath = "//a:r//a:t")
}
\arguments{
\item{file}{XML file from a PPTX}

\item{collapse_text}{should text be collapsed by spaces?}

\item{xpath}{\code{xpath} to pass to \code{\link[xml2:xml_find_all]{xml2::xml_find_all()}}}
}
\value{
A character vector
}
\description{
Extract Notes from XML
}
